+++ internal/common/conversion/misc.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/spf13/cast"
 )
 
@@ -29,19 +28,4 @@ func ValRegion(reg any, opt ...string) (string, error) {
 	}
 
 	return strings.ReplaceAll(region, "-", "_"), nil
-}
-
-func ExpandStringList(list []any) (res []string) {
-	for _, v := range list {
-		res = append(res, v.(string))
-	}
-	return
-}
-
-func ExpandStringListFromSetSchema(set *schema.Set) []string {
-	res := make([]string, set.Len())
-	for i, v := range set.List() {
-		res[i] = v.(string)
-	}
-	return res
 }
